+++ src/main/java/roomescape/dto/ReservationResponse.java
@@ -2,7 +2,7 @@ package roomescape.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
-import roomescape.entity.Reservation;
+import roomescape.model.Reservation;
 
 public record ReservationResponse(Long id,
                                   String name,
