+++ src/main/java/roomescape/model/Reservation.java
@@ -1,6 +1,7 @@
-package roomescape.entity;
+package roomescape.model;
 
 import java.time.LocalDate;
+import roomescape.entity.ReservationTime;
 
 public class Reservation {
